backend/python/requirements.dev.txt
annotated-types==0.6.0
openai
pytube3
ollama
youtube-transcript-api
stripe
supabase
black
fuzzywuzzy
moviepy
dlib
face_recognition
scipy
python-dotenv
opencv-python-headless
pydub
assemblyai
